requirements.txt
streamlit>=1.35.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python-headless>=4.9.0
pandas>=2.2.0
matplotlib>=3.8.0
Pillow>=10.0.0
numpy>=1.26.0

